+++ src/Heretic.Roguelike/Utils/IContentPrinter.cs
@@ -16,7 +16,7 @@ public interface IContentPrinter<T, TK> where TK : ICell<T>
 
     void DrawItemAtPosition(IList<TK> cells, Vector position, T item);
     
-    void DrawDashboard(IList<TK> cells, Player<T> creature, ushort currentFloor);
+    void DrawDashboard(IList<TK> cells, Player<T> creature, int currentFloor);
     
     void DrawMessage(IList<TK> cells, string message);
     void ClearMessage(IList<TK> cells);
